# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mystics ERP API — Sales & CRM, Project Management, Inventory & Warehouse Management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: getProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: pmOwnerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}/dashboard:
    get:
      operationId: getProjectDashboard
      tags: [projects]
      summary: Single-project KPI dashboard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDashboard"
  /projects/{id}/budget-vs-actual:
    get:
      operationId: getProjectBudgetVsActual
      tags: [projects]
      summary: Budget vs actual for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Budget variance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetVsActual"
...
  /projects/{id}/activities:
    get:
      operationId: getProjectActivities
      tags: [activities]
      summary: Get activities / Gantt data for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Add activity to a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
  /activities/{id}:
    patch:
      operationId: updateActivity
      tags: [activities]
      summary: Update an activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"

  # ── MATERIAL REQUESTS ──────────────────────────────────────────────────────
...
  /dprs:
    get:
      operationId: getDPRs
      tags: [dprs]
      summary: List DPRs for a project
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: DPRs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DPR"
    post:
      operationId: createDPR
      tags: [dprs]
      summary: Submit a daily progress report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DPRInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DPR"

  # ── PAYMENT MILESTONES ─────────────────────────────────────────────────────
  /projects/{id}/payment-milestones:
    get:
      operationId: getPaymentMilestones
      tags: [milestones]
      summary: List payment milestones for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentMilestone"
    post:
      operationId: createPaymentMilestone
      tags: [milestones]
      summary: Create a payment milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentMilestoneInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentMilestone"
  /payment-milestones/{id}/trigger:
    post:
      operationId: triggerPaymentMilestone
      tags: [milestones]
      summary: Mark milestone complete and fire billing event to CRM
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentMilestone"

  # ── EXPENSES ───────────────────────────────────────────────────────────────
  /expenses:
    get:
      operationId: getExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Submit an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}/approve:
    post:
      operationId: approveExpense
      tags: [expenses]
      summary: Approve or reject an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"

  # ── BUDGETS ────────────────────────────────────────────────────────────────
  /budgets:
    get:
      operationId: getBudgets
      tags: [budgets]
      summary: List budget lines
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create a budget line
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"

  # ── CONTRACTORS ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApprovalInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
        remarks:
          type: string
...
    Project:
      type: object
      required: [id, name, status, createdAt]
      properties:
        id:
          type: integer
        clientPoId:
          type: ["integer", "null"]
        name:
          type: string
        siteLocation:
          type: ["string", "null"]
        pmOwnerId:
          type: ["integer", "null"]
        pmOwnerName:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        plannedEnd:
          type: ["string", "null"]
        status:
          type: string
        parentProjectId:
          type: ["integer", "null"]
        contractValue:
          type: ["number", "null"]
        percentComplete:
          type: ["number", "null"]
        createdAt:
          type: string
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        clientPoId:
          type: integer
        siteLocation:
          type: string
        pmOwnerId:
          type: integer
        startDate:
          type: string
        plannedEnd:
          type: string
        contractValue:
          type: number
        parentProjectId:
          type: integer
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        siteLocation:
          type: string
        pmOwnerId:
          type: integer
        startDate:
          type: string
        plannedEnd:
          type: string
        status:
          type: string
        percentComplete:
          type: number
    ProjectDashboard:
      type: object
      properties:
        project:
          $ref: "#/components/schemas/Project"
        budgetSummary:
          $ref: "#/components/schemas/BudgetVsActual"
        activitiesCount:
          type: integer
        openMRsCount:
          type: integer
        pendingPOsCount:
          type: integer
        openEscalationsCount:
          type: integer
        lastDPR:
          $ref: "#/components/schemas/DPR"
        upcomingMilestones:
          type: array
          items:
            $ref: "#/components/schemas/PaymentMilestone"
    BudgetVsActual:
      type: object
      properties:
        projectId:
          type: integer
        lines:
          type: array
          items:
            type: object
            properties:
              costHead:
                type: string
              budgeted:
                type: number
              committed:
                type: number
              actual:
                type: number
              variance:
                type: number
            required: [costHead, budgeted, committed, actual, variance]
        totalBudgeted:
          type: number
        totalCommitted:
          type: number
        totalActual:
          type: number
        totalVariance:
          type: number
      required: [projectId, lines, totalBudgeted, totalCommitted, totalActual, totalVariance]

    # ── ACTIVITIES ────────────────────────────────────────────────────────────
    Activity:
      type: object
      required: [id, projectId, name, status]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        wbsCode:
          type: ["string", "null"]
        name:
          type: string
        plannedStart:
          type: ["string", "null"]
        plannedEnd:
          type: ["string", "null"]
        actualStart:
          type: ["string", "null"]
        actualEnd:
          type: ["string", "null"]
        dependencyIds:
          type: array
          items:
            type: integer
        percentComplete:
          type: number
        status:
          type: string
    ActivityInput:
      type: object
      required: [name]
      properties:
        wbsCode:
          type: string
        name:
          type: string
        plannedStart:
          type: string
        plannedEnd:
          type: string
        dependencyIds:
          type: array
          items:
            type: integer
        status:
          type: string
    ActivityUpdate:
      type: object
      properties:
        name:
          type: string
        plannedStart:
          type: string
        plannedEnd:
          type: string
        actualStart:
          type: string
        actualEnd:
          type: string
        percentComplete:
          type: number
        status:
          type: string

    # ── MATERIAL REQUESTS ─────────────────────────────────────────────────────
...
    DPR:
      type: object
      required: [id, projectId, reportDate, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        reportDate:
          type: string
        submittedBy:
          type: ["integer", "null"]
        submittedByName:
          type: ["string", "null"]
        workSummary:
          type: ["string", "null"]
        manpowerCount:
          type: ["integer", "null"]
        weather:
          type: ["string", "null"]
        percentComplete:
          type: ["number", "null"]
        photos:
          type: array
          items:
            type: string
        createdAt:
          type: string
    DPRInput:
      type: object
      required: [projectId, reportDate]
      properties:
        projectId:
          type: integer
        reportDate:
          type: string
        workSummary:
          type: string
        manpowerCount:
          type: integer
        weather:
          type: string
        percentComplete:
          type: number
        photos:
          type: array
          items:
            type: string

    # ── PAYMENT MILESTONES ────────────────────────────────────────────────────
    PaymentMilestone:
      type: object
      required: [id, projectId, milestoneName, amount, status]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        milestoneName:
          type: string
        triggerCondition:
          type: ["string", "null"]
        amount:
          type: number
        dueDate:
          type: ["string", "null"]
        status:
          type: string
        invoiceRef:
          type: ["integer", "null"]
    PaymentMilestoneInput:
      type: object
      required: [milestoneName, amount]
      properties:
        milestoneName:
          type: string
        triggerCondition:
          type: string
        amount:
          type: number
        dueDate:
          type: string

    # ── EXPENSES ──────────────────────────────────────────────────────────────
    Expense:
      type: object
      required: [id, projectId, category, amount, approvalStatus, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        category:
          type: string
        amount:
          type: number
        incurredBy:
          type: ["integer", "null"]
        incurredByName:
          type: ["string", "null"]
        date:
          type: string
        receiptUrl:
          type: ["string", "null"]
        approvalStatus:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required: [projectId, category, amount, date]
      properties:
        projectId:
          type: integer
        category:
          type: string
        amount:
          type: number
        date:
          type: string
        receiptUrl:
          type: string
        notes:
          type: string

    # ── BUDGETS ───────────────────────────────────────────────────────────────
    Budget:
      type: object
      required: [id, projectId, costHead, budgetedAmount]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        costHead:
          type: string
        budgetedAmount:
          type: number
        committedAmount:
          type: number
        actualAmount:
          type: number
        revisionNo:
          type: integer
    BudgetInput:
      type: object
      required: [projectId, costHead, budgetedAmount]
      properties:
        projectId:
          type: integer
        costHead:
          type: string
        budgetedAmount:
          type: number

    # ── CONTRACTORS ───────────────────────────────────────────────────────────
...