# lib/api-zod/src/generated/types/dPRInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Mystics ERP API — Sales & CRM, Project Management, Inventory & Warehouse Management
 * OpenAPI spec version: 0.1.0
 */

export interface DPRInput {
  projectId: number;
  reportDate: string;
  workSummary?: string;
  manpowerCount?: number;
  weather?: string;
  percentComplete?: number;
  photos?: string[];
}